Function ceph_rust::json::json_find

pub fn json_find(json_data: JsonData, keys: &[&str]) -> Option<JsonData>

Looks for the parent object first and then the 'child' object. If the parent object is None then it only looks for the 'child' object. The parent object is used for situations where there may be 'child' objects with the same name.