Function ceph_rust::utils::run_cli

pub fn run_cli(cmd_line: &str) -> Result<Output>

run_cli - pass in a String of a normal command line

The function will split the options into words to supply to the low_level std::process::Command which returns Result<(Output)>

Example

run_cli("ps aux");